refactor api routes and add Policy Gateway for Controller BookCopyController

+++ app/Http/Controllers/BookCopyController.php
@@ -6,10 +6,15 @@ use App\Http\Requests\BookCopyRequest;
 use App\Http\Requests\BookCopyUpdateRequest;
 use App\Models\BookCopy;
 use App\Services\BookCopyService;
-use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Support\Facades\Gate;
 
 class BookCopyController extends Controller
 {
+    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+
     private BookCopyService $bookCopyService;
 
     public function __construct(BookCopyService $bookCopyService)
@@ -24,16 +29,25 @@ class BookCopyController extends Controller
 
     public function store(BookCopyRequest $request)
     {
+        if (!Gate::allows('store', BookCopy::class)) {
+            abort(403, 'شما دسترسی به این بخش ندارید.');
+        }
         return $this->bookCopyService->store($request);
     }
 
-    public function update(BookCopyUpdateRequest $request , BookCopy $bookCopy)
+    public function update(BookCopyUpdateRequest $request, BookCopy $bookCopy)
     {
-        return $this->bookCopyService->update($request , $bookCopy);
+        if (!Gate::allows('update', BookCopy::class)) {
+            abort(403, 'شما دسترسی به این بخش ندارید.');
+        }
+        return $this->bookCopyService->update($request, $bookCopy);
     }
 
     public function destroy(BookCopy $bookCopy)
     {
+        if (!Gate::allows('delete', BookCopy::class)) {
+            abort(403, 'شما دسترسی به این بخش ندارید.');
+        }
         return $this->bookCopyService->delete($bookCopy);
     }
 }
